moves fixtures to a separate file
