--- network/p2p/unittest/fixture.go
@@ -1,1 +1,0 @@
-package p2p
